+++ icevision/models/rcnn/faster_rcnn/__init__.py
@@ -4,6 +4,7 @@ from icevision.models.rcnn.loss_fn import *
 from icevision.models.rcnn.faster_rcnn.dataloaders import *
 from icevision.models.rcnn.faster_rcnn.model import *
 from icevision.models.rcnn.faster_rcnn.prediction import *
+from icevision.models.rcnn.faster_rcnn.show_results import *
 
 
 # Soft dependencies