+++ tests/models/rcnn/faster_rcnn/test_show_results.py
@@ -1,0 +1,30 @@
+import pytest
+from icevision.all import *
+
+
+@pytest.fixture(scope="session")
+def fake_faster_rcnn_model():
+    class FakeFasterRCNNModel(nn.Module):
+        def __init__(self):
+            super().__init__()
+            # hack for the function `model_device` to work
+            self.layer = nn.Linear(1, 1)
+
+        def forward(self, *args, **kwargs):
+            return [
+                {
+                    "scores": tensor([0.8, 0.9]),
+                    "labels": tensor([1, 2]),
+                    "boxes": tensor([[10, 10, 50, 50], [20, 20, 40, 40]]),
+                }
+            ]
+
+    return FakeFasterRCNNModel()
+
+
+def test_faster_rcnn_show_results(fake_faster_rcnn_model, fridge_ds):
+    train_ds, valid_ds = fridge_ds
+
+    faster_rcnn.show_results(
+        model=fake_faster_rcnn_model, dataset=valid_ds, num_samples=1, ncols=1
+    )