+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ## Added
+- `show_results` function for each model
 
 ## Changed
 - Renamed package from `mantisshrimp` to `icevision`